refactor get api, rendering with id is working now
